Gets user input such as name, age, height and weight. Outputs name, age and a calculated bmi from weight and height

+++ basics/input/data_types.py
@@ -2,11 +2,11 @@
 # Gets the users name
 name = input("What is your name Human?")
 # Gets the users age
-age = input("How old are you (in years)?")
+age = int(input("How old are you (in years)?"))
 # Gets the users height and allocates it as a float point variable
 height = float(input("How tall are you (in meters)?"))
-# Gets the users height and allocates it as a integer variable
-weight = int(input("How much do you weigh (in kilograms)?"))
+# Gets the users weight and allocates it as a integer variable
+weight = float(input("How much do you weigh (in kilograms)?"))
 # Calculates the users bmi
 bmi = str(round(weight / height, 2))
 # Output
